api/requirements.txt
setuptools>=70.0.0
wheel>=0.42.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pymongo>=4.6.0
motor>=3.4.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
pydantic>=2.6.0
google-generativeai>=0.4.0
chromadb>=0.4.24
sentence-transformers>=2.5.0
requests>=2.31.0
beautifulsoup4>=4.12.3
numpy>=1.26.0
pandas>=2.1.0
aiofiles>=23.2.1
httpx>=0.26.0
tenacity>=8.2.3
certifi>=2023.11.17
urllib3>=2.0.0
cryptography>=41.0.0
yfinance>=0.2.66
email-validator>=2.0.0
bcrypt==4.2.1
